Rename internal package "utils" to "honestnft_utils"
