+++ honestnft_utils/chain.py
@@ -8,8 +8,8 @@ from multicall import Call, Multicall
 from web3 import Web3
 from web3.exceptions import ContractLogicError
 
-from utils import config
-from utils import ipfs
+from honestnft_utils import config
+from honestnft_utils import ipfs
 
 
 def get_contract_abi(address, blockchain="ethereum"):
