+++ honestnft_utils/opensea.py
@@ -2,7 +2,7 @@ import time
 
 import requests
 
-from utils import config
+from honestnft_utils import config
 
 SLEEP = 5
