+++ metadata/pulling.py
@@ -10,9 +10,9 @@ import pandas as pd
 import requests
 from web3.exceptions import ContractLogicError
 
-from utils import chain
-from utils import config
-from utils import ipfs
+from honestnft_utils import chain
+from honestnft_utils import config
+from honestnft_utils import ipfs
 
 
 """
